method edit + new final static int for deck size.
Pujara also told us to use a final static int SIZE for the deck size, so I added it and set it to 28. That might call for some changes in later lines of code.

I edited the getFileName method to include reading the file and storing the data in the private static deck array. 
   - it also uses a try-catch now instead of canRead() in order to check whether or not the input is a valid file name. 

Also fixed some indenting errors